requirements.txt
scapy == 2.4.2
opencv_python
requests
colorama
pyautogui
pysimplegui
pillow
pywin32